# backend/src/routes/caption/index.ts
import { Router, type IRouter } from "express";
import { openai } from "@workspace/integrations-openai-ai-server";
import { requireAuth, requirePlanOrTrial } from "../../middlewares/planMiddleware";
import { LANGUAGE_INSTRUCTIONS } from "../../lib/languages";

const router: IRouter = Router();

const PLATFORM_CONTEXT: Record<string, string> = {
  Instagram: "Instagram caption (max 2200 chars). Uses line breaks for scanability. 3-5 relevant hashtags at end. First line must be the hook — only 1-2 lines show before 'more'. Emojis enhance personality, never overuse. CTA in last line.",
  Twitter: "Twitter/X thread or single tweet (280 chars). Punchy, fast-moving, opinionated. Every word earns its place. Hook triggers immediate engagement or controversy.",
  LinkedIn: "LinkedIn post. Professional but human. Storytelling format: setup → tension → insight → lesson. No hashtag spam (max 3). Ends with a question or clear CTA. Line breaks create breathing room.",
  YouTube: "YouTube video description (SEO-optimized). First 125 chars visible before 'more' — must hook AND include primary keyword. Timestamps if relevant. Keywords naturally woven in. End with subscribe/subscribe CTA.",
  "Blog/Article": "Blog post intro/meta description. SEO-friendly, grabs attention in first 2 sentences, makes the reader feel they MUST read on.",
  General: "Universal social media caption. Engaging, clear, conversation-starting. Focus on the emotional payoff for the reader.",
};

import { generateContent } from "../../services/ai-engine";

router.post("/caption/enhance", requireAuth, requirePlanOrTrial("caption"), async (req: any, res): Promise<void> => {
  const {
    originalCaption,
    platform = "Instagram",
    goal = "increase engagement",
    niche = "General",
    improvementFocus = "all",
    language = "English"
  } = req.body;

  if (!originalCaption?.trim()) {
    res.status(400).json({ error: "Original caption is required" });
    return;
  }

  const platformContext = PLATFORM_CONTEXT[platform] || PLATFORM_CONTEXT["General"];

  const systemPrompt = `You are a viral content strategist. Optimize this ${platform} caption.
RULES: No "I", no generic openers, no exclamation spam.
CONTEXT: ${platformContext}.`;

  const userPrompt = `Analyze and enhance this caption. Goal: ${goal}. Niche: ${niche}.
ORIGINAL: "${originalCaption}"
FOCUS: ${improvementFocus}

Return ONLY this JSON: {
  "diagnosis": {"mainIssue": "string", "strengths": [], "weaknesses": []},
  "fullRewrite": {"caption": "string", "changesMade": [], "whyItWorks": "string"},
  "microEdit": {"caption": "string", "changesMade": []},
  "hookScore": {"original": 0, "rewrite": 0, "explanation": "string"}
}`;

  try {
    const rawContentObj = await generateContent({
      messages: [
        { role: "system", content: systemPrompt },
        { role: "user", content: userPrompt }
      ],
      userPlan: req.user?.planType || "free",
      userId: req.userId,
      language,
      maxTokens: 2500,
    });

    const raw = rawContentObj.choices[0]?.message?.content ?? "{}";
    let parsed;
    try {
      const match = raw.match(/\{[\s\S]*\}/);
      parsed = JSON.parse(match ? match[0] : raw);
    } catch {
      throw new Error("Failed to parse caption response.");
    }
    res.json(parsed);
  } catch (err: any) {
    console.error("Caption enhance error:", err);
    res.status(503).json({ error: "AI temporarily unavailable." });
  }
});

export default router;