# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * Grow Flow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GenerateContentBodyContentType = typeof GenerateContentBodyContentType[keyof typeof GenerateContentBodyContentType];


export const GenerateContentBodyContentType = {
  Educational: 'Educational',
  Story: 'Story',
  Viral: 'Viral',
} as const;

export type GenerateContentBodyTone = typeof GenerateContentBodyTone[keyof typeof GenerateContentBodyTone];


export const GenerateContentBodyTone = {
  Casual: 'Casual',
  Professional: 'Professional',
  Aggressive: 'Aggressive',
} as const;

/**
 * Optional - language of the generated content
 */
export type GenerateContentBodyLanguage = typeof GenerateContentBodyLanguage[keyof typeof GenerateContentBodyLanguage];


export const GenerateContentBodyLanguage = {
  English: 'English',
  Hindi: 'Hindi',
  Hinglish: 'Hinglish',
  Bengali: 'Bengali',
  Gujarati: 'Gujarati',
  Kannada: 'Kannada',
  Malayalam: 'Malayalam',
  Marathi: 'Marathi',
  Punjabi: 'Punjabi',
  Odia: 'Odia',
  Assamese: 'Assamese',
  Tamil: 'Tamil',
  Telugu: 'Telugu',
  Urdu: 'Urdu',
  Bhojpuri: 'Bhojpuri',
  Spanish: 'Spanish',
  French: 'French',
  German: 'German',
} as const;

export interface GenerateContentBody {
  /** The content idea to generate from */
  idea: string;
  contentType: GenerateContentBodyContentType;
  tone: GenerateContentBodyTone;
  /** Optional - language of the generated content */
  language?: GenerateContentBodyLanguage;
}

export type GenerateVariationsBodyContentType = typeof GenerateVariationsBodyContentType[keyof typeof GenerateVariationsBodyContentType];


export const GenerateVariationsBodyContentType = {
  Educational: 'Educational',
  Story: 'Story',
  Viral: 'Viral',
} as const;

export type GenerateVariationsBodyTone = typeof GenerateVariationsBodyTone[keyof typeof GenerateVariationsBodyTone];


export const GenerateVariationsBodyTone = {
  Casual: 'Casual',
  Professional: 'Professional',
  Aggressive: 'Aggressive',
} as const;

/**
 * Optional - language of the generated content
 */
export type GenerateVariationsBodyLanguage = typeof GenerateVariationsBodyLanguage[keyof typeof GenerateVariationsBodyLanguage];


export const GenerateVariationsBodyLanguage = {
  English: 'English',
  Hindi: 'Hindi',
  Hinglish: 'Hinglish',
  Bengali: 'Bengali',
  Gujarati: 'Gujarati',
  Kannada: 'Kannada',
  Malayalam: 'Malayalam',
  Marathi: 'Marathi',
  Punjabi: 'Punjabi',
  Odia: 'Odia',
  Assamese: 'Assamese',
  Tamil: 'Tamil',
  Telugu: 'Telugu',
  Urdu: 'Urdu',
  Bhojpuri: 'Bhojpuri',
  Spanish: 'Spanish',
  French: 'French',
  German: 'German',
} as const;

/**
 * Optional - specific platform to vary
 */
export type GenerateVariationsBodyPlatform = typeof GenerateVariationsBodyPlatform[keyof typeof GenerateVariationsBodyPlatform];


export const GenerateVariationsBodyPlatform = {
  instagram: 'instagram',
  youtube: 'youtube',
  twitter: 'twitter',
  linkedin: 'linkedin',
} as const;

export interface GenerateVariationsBody {
  /** ID of the history item to generate variations for */
  historyId?: number;
  idea: string;
  contentType: GenerateVariationsBodyContentType;
  tone: GenerateVariationsBodyTone;
  /** Optional - language of the generated content */
  language?: GenerateVariationsBodyLanguage;
  /** Optional - specific platform to vary */
  platform?: GenerateVariationsBodyPlatform;
}

export type PlatformContentInstagram = {
  caption: string;
  hashtags: string;
};

export type PlatformContentYoutube = {
  hook: string;
  script: string;
};

export type PlatformContentTwitter = {
  tweets: string[];
};

export type PlatformContentLinkedin = {
  post: string;
};

export interface PlatformContent {
  instagram: PlatformContentInstagram;
  youtube: PlatformContentYoutube;
  twitter: PlatformContentTwitter;
  linkedin: PlatformContentLinkedin;
}

export interface GeneratedContent {
  id: number;
  idea: string;
  contentType: string;
  tone: string;
  content: PlatformContent;
  createdAt: string;
}

export interface HistoryItem {
  id: number;
  idea: string;
  contentType: string;
  tone: string;
  content: PlatformContent;
  createdAt: string;
}

export interface HistoryResponse {
  items: HistoryItem[];
  total: number;
}

export type GenerateHooksBodyTone = typeof GenerateHooksBodyTone[keyof typeof GenerateHooksBodyTone];


export const GenerateHooksBodyTone = {
  Casual: 'Casual',
  Professional: 'Professional',
  Aggressive: 'Aggressive',
} as const;

export interface GenerateHooksBody {
  /** The topic to generate hooks for */
  topic: string;
  tone: GenerateHooksBodyTone;
}

export interface GeneratedHooks {
  hooks: string[];
}

export type ContentStatsPlatformBreakdown = {
  instagram: number;
  youtube: number;
  twitter: number;
  linkedin: number;
};

export interface ContentStats {
  totalGenerations: number;
  thisWeek: number;
  topContentType: string;
  topTone: string;
  platformBreakdown: ContentStatsPlatformBreakdown;
}

export interface ImproveCompetitorContentBody {
  competitorContent: string;
}

export type ImproveCompetitorContentResultMonetization = {
  ctaIdeas: string[];
  affiliateSuggestions: string[];
  howToEarn: string[];
};

export interface ImproveCompetitorContentResult {
  competitorWeaknesses: string[];
  missedAngle: string;
  improvedVersion: string;
  strongerHook: string;
  newAngle: string;
  monetization: ImproveCompetitorContentResultMonetization;
}

export type GetContentHistoryParams = {
limit?: number;
offset?: number;
};

export type DeleteHistoryItem200 = {
  success: boolean;
};